requirements.txt
click==6.6
Flask==0.11
Flask-SocketIO==2.3
itsdangerous==0.24
Jinja2==2.8
kafka-python==1.2.0
MacroPy==1.0.3
MarkupSafe==0.23
python-engineio==0.9.1
python-socketio==1.3
requests==2.10.0
six==1.10.0
Werkzeug==0.11.10
wheel==0.24.0
